replace all clis for a single one
